Comparison between signed and unsigned integrals has been omitted.
git-svn-id: svn+ssh://cv077/home/ueshiba/svnroot/trunk/src@2449 3ef1b8b6-fb57-41bf-88d9-9a5bd220bf38

+++ featureCmds.cc
@@ -208,8 +208,9 @@ refreshFeatureCmds(const IIDCCamera& camera, CmdPane& cmdPane)
 	
 	if (getFeature(camera, id, val, fval))
 	{
-	    if ((id >= IIDCCamera::BRIGHTNESS + IIDCCAMERA_OFFSET_ONOFF) ||
-		(id == IIDCCamera::TRIGGER_MODE))
+	    if ((u_int(id) >= IIDCCamera::BRIGHTNESS
+			    + IIDCCAMERA_OFFSET_ONOFF) ||
+		(u_int(id) == IIDCCamera::TRIGGER_MODE))
 		cmdPane.setValue(id, val);
 	  // White bal.(V/R)は id == WHITE_BALANCE の時に一緒にrefreshされる
 	    else if (id != IIDCCamera::WHITE_BALANCE + IIDCCAMERA_OFFSET_VR)
